Devhelp: add tests
